requirements.txt
apscheduler>=3.3.0
aws-requests-auth>=0.3.0
blist>=1.3.6
boto3>=1.4.4
cffi>=1.11.5
configparser>=3.5.0
croniter>=0.3.16
dataclasses
elasticsearch>=7.0.0
envparse>=0.2.0
exotel>=0.1.3
jira>=2.0.0
jsonschema>=3.0.2
mock>=2.0.0
prison>=0.1.2
py-zabbix==1.1.3
PyStaticConfiguration>=0.10.3
python-dateutil>=2.6.0,<2.7.0
Python-Deprecated
python-magic>=0.4.15
pytz
PyYAML>=5.1
requests>=2.22.0
six
stomp.py>=4.1.17
texttable>=0.8.8
twilio==6.0.0
